+++ R/functions.R
@@ -171,6 +171,8 @@ plot_smooths <- function(model, time_series, comparison, facet_terms = NULL, con
 #' @importFrom magrittr "%>%"
 #' @export
 plot_gamsd <- function(model, view, comparison, conditions = NULL, rm_re = FALSE, bw = FALSE) {
+    .Deprecated("plot_smooth", msg = "'plot_gamsd' will be deprecated, use 'plot_smooths'. (Plotting with the difference smooth is not supported yet.)\n")
+
     diff.df <- itsadug::plot_diff(
         model,
         view = view,