+++ NEWS.md
@@ -1,5 +1,8 @@
 # Change Log
 
+## Unreleased
+### Fixed
+- typo in documentation of `create_event_start()`
 ## 1.2.0 - 2017-12-07
 ### Added
 - `rm_re` parameter for removing random effects in `plot_gamsd`