+++ NEWS.md
@@ -3,6 +3,7 @@
 ## Unreleased
 ### Fixed
 - typo in documentation of `create_event_start()`
+- note about `lag` on check
 ## 1.2.0 - 2017-12-07
 ### Added
 - `rm_re` parameter for removing random effects in `plot_gamsd`