+++ NEWS.md
@@ -3,7 +3,10 @@
 ## Unreleased
 ### Fixed
 - typo in documentation of `create_event_start()`
+- double call of `itsadug` in DESCRIPTION
 - note about `lag` on check
+- note about non-imported `itsadug`
+
 ## 1.2.0 - 2017-12-07
 ### Added
 - `rm_re` parameter for removing random effects in `plot_gamsd`