+++ NEWS.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 ### Added
+- `plot_smooths` for plotting smooths
 - `plot_gami()` for plotting interactions of generalised additive models
 - `plot_gami_all()` for plotting all the levels of an interaction of generalised additive models