+++ NEWS.md
@@ -12,7 +12,7 @@
 - note about non-imported `itsadug`
 
 ### Deprecated
-- `plot_gamsd()`: use `plot_smooths` instead. Plotting both smooths and difference smooth is not supported yet.
+- `plot_gamsd()` will be deprecated: use `plot_smooths` instead. Plotting both smooths and difference smooth is not supported yet.
 
 ## 1.2.0 - 2017-12-07
 ### Added